requirements.txt
cupy-cuda13x
numpy
pandas
matplotlib
tqdm
scikit-learn
tensorflow
keras